delete

+++ anime/app/routers/json_data.py
@@ -1,4 +1,4 @@
-
+import os
 from fastapi import APIRouter
 from app.db.deps import db_dependency
 from app.lib import serializes_delete
@@ -12,7 +12,7 @@ async def write_animes(db: db_dependency) -> dict:
     anime = db.animes.find()
     anime_list = [serializes_delete(a) async for a in anime]
     
-    with open("data/animes.json", "w", encoding="utf-8") as f:
+    with open("data/anime.json", "w", encoding="utf-8") as f:
         json.dump(anime_list, f, ensure_ascii=False, indent=2)
     
     return {"message": "Animes saved to file"}
@@ -20,9 +20,12 @@ async def write_animes(db: db_dependency) -> dict:
 @router.get("/read")
 async def read_animes(db: db_dependency) -> dict:
     
+    if not os.path.exists("data/anime.json"):
+        raise HTTPException(status_code=404, detail="File data/anime.json not found")
+
     await db.animes.delete_many({})
     
-    with open("data/animes.json", "r", encoding="utf-8") as f:
+    with open("data/anime.json", "r", encoding="utf-8") as f:
         anime_list = json.load(f)
 
     for anime in anime_list:
